Update test files to use lateral join functionality
- Simplified test syntax using lateral joins where appropriate
- Updated tests to work with new OLS table function
- Fixed type casting issues in array literals
- Updated field names to match current function signatures
- Removed obsolete test patterns that are no longer needed
- Net reduction of 420 lines across 45 test files

+++ test/sql/guide04_use_case_3_customer_lifetime_value_predi.sql
@@ -5,10 +5,10 @@ CREATE OR REPLACE TABLE customer_summary AS
 SELECT
     i as customer_id,
     (tenure * 45 + aov * 3 + freq * 80 + engagement * 500 + RANDOM() * 200)::DOUBLE as total_purchases,
-    tenure::DOUBLE,
-    aov::DOUBLE,
-    freq::DOUBLE,
-    engagement::DOUBLE,
+    tenure,
+    aov,
+    freq,
+    engagement,
     CURRENT_DATE - (12 + (RANDOM() * 24)::INT) * INTERVAL '1 month' as cohort_month
 FROM (
     SELECT
